Add CI workflow and fix lint/type errors for branch protection
Add GitHub Actions CI matching KITT's strategy: test matrix across
Python 3.10-3.13 on Ubuntu + 3.12 on macOS, lint job with ruff,
black, and mypy. Fix pre-existing lint violations (unused imports,
ambiguous variable names, unused assignments) and type errors
(incorrect type hints, missing stubs).

+++ src/devon/utils/size_parser.py
@@ -30,7 +30,7 @@ def parse_size(size_str: str) -> Optional[int]:
     return int(value * multipliers[unit])
 
 
-def format_bytes(bytes_val: int) -> str:
+def format_bytes(bytes_val: float) -> str:
     """Format bytes to human-readable string."""
     if bytes_val < 0:
         return "0B"
